requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
Faker==1.0.2
idna==2.8
pyhunter==0.3
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
text-unidecode==1.2
urllib3==1.24.1
